# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Society Events API",
    "description": "REST API for the Society Events platform — GM Global Techies Town, Bengaluru.\n\nAll endpoints require a Bearer JWT issued by Keycloak **except** `/payments/webhook`.\n\n**Roles:**\n- `resident` — browse, register, pay\n- `committee_member` — create/manage events, send announcements\n- `admin` — full access including user management and reports\n- `security_guard` — mark attendance via QR scan\n\n**Obtain a token:**\n```\nPOST http://localhost:8080/realms/society-events/protocol/openid-connect/token\nContent-Type: application/x-www-form-urlencoded\n\nclient_id=society-frontend&grant_type=password&username=rajesh.iyer@pvh-blr.in&password=Test@1234\n```",
    "version": "1.0.0",
    "contact": {
      "name": "Society Admin",
      "email": "admin@pvh-blr.in"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Local API Gateway"
    }
  ],
  "security": [
    { "bearerAuth": [] }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "UUID": {
        "type": "string",
        "format": "uuid",
        "example": "51100000-0000-0000-0000-000000000001"
      },
      "Error": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {
          "code":    { "type": "string", "example": "RESOURCE_NOT_FOUND" },
          "message": { "type": "string", "example": "Event not found" }
        }
      },
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "page":  { "type": "integer", "example": 1 },
          "limit": { "type": "integer", "example": 20 },
          "total": { "type": "integer", "example": 42 }
        }
      },
      "Apartment": {
        "type": "object",
        "properties": {
          "id":          { "$ref": "#/components/schemas/UUID" },
          "society_id":  { "$ref": "#/components/schemas/UUID" },
          "block":       { "type": "string", "example": "A" },
          "unit_number": { "type": "string", "example": "101" },
          "type":        { "type": "string", "enum": ["1BHK", "2BHK", "3BHK", "Villa"] }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id":           { "$ref": "#/components/schemas/UUID" },
          "name":         { "type": "string", "example": "Rajesh Iyer" },
          "email":        { "type": "string", "format": "email", "example": "rajesh.iyer@pvh-blr.in" },
          "phone":        { "type": "string", "example": "+91-98450-11111" },
          "role":         { "type": "string", "enum": ["admin", "committee_member", "resident", "security_guard"] },
          "apartment_id": { "$ref": "#/components/schemas/UUID" },
          "apartment":    { "$ref": "#/components/schemas/Apartment" },
          "is_active":    { "type": "boolean" },
          "created_at":   { "type": "string", "format": "date-time" }
        }
      },
      "UpdateProfileRequest": {
        "type": "object",
        "properties": {
          "name":  { "type": "string" },
          "phone": { "type": "string" }
        }
      },
      "UpdateRoleRequest": {
        "type": "object",
        "required": ["role"],
        "properties": {
          "role": { "type": "string", "enum": ["admin", "committee_member", "resident", "security_guard"] }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid JWT",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Forbidden": {
        "description": "Insufficient role",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
    },
    "parameters": {
      "PageParam": {
        "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 }
      },
      "LimitParam": {
        "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20, "maximum": 100 }
      }
    }
  },
  "paths": {
    "/users/me": {
      "get": {
        "tags": ["Users"],
        "summary": "Get my profile",
        "responses": {
          "200": { "description": "Current user profile", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "put": {
        "tags": ["Users"],
        "summary": "Update my profile (name, phone)",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateProfileRequest" } } } },
        "responses": {
          "200": { "description": "Updated profile", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "List all users — admin only",
        "parameters": [
          { "$ref": "#/components/parameters/PageParam" },
          { "$ref": "#/components/parameters/LimitParam" },
          { "name": "role", "in": "query", "schema": { "type": "string", "enum": ["admin", "committee_member", "resident", "security_guard"] } },
          { "name": "q", "in": "query", "description": "Search by name or email", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Paginated user list",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "meta": { "$ref": "#/components/schemas/PaginationMeta" }, "data": { "type": "array", "items": { "$ref": "#/components/schemas/User" } } } } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/users/{id}": {
      "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/UUID" } }],
      "get": {
        "tags": ["Users"],
        "summary": "Get user by ID — admin only",
        "responses": {
          "200": { "description": "User", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/users/{id}/role": {
      "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/UUID" } }],
      "put": {
        "tags": ["Users"],
        "summary": "Update user role — admin only",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateRoleRequest" } } } },
        "responses": {
          "200": { "description": "Updated user", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  }
}